+++ src/faim_hcs/hcs/cellvoyager/StackAcquisition.py
@@ -1,12 +1,10 @@
-import re
 from pathlib import Path
 from typing import Optional, Union
 
-from faim_hcs.hcs.acquisition import TileAlignmentOptions
-from faim_hcs.hcs.imagexpress import ImageXpressPlateAcquisition
+from faim_hcs.hcs.acquisition import PlateAcquisition, TileAlignmentOptions
 
 
-class SinglePlaneAcquisition(ImageXpressPlateAcquisition):
+class StackAcquistion(PlateAcquisition):
     def __init__(
         self,
         acquisition_dir: Union[Path, str],
@@ -20,14 +18,3 @@ class SinglePlaneAcquisition(ImageXpressPlateAcquisition):
             background_correction_matrices=background_correction_matrices,
             illumination_correction_matrices=illumination_correction_matrices,
         )
-
-    def _get_root_re(self) -> re.Pattern:
-        return re.compile(r".*[\/\\](?P<date>\d{4}-\d{2}-\d{2})[\/\\](?P<acq_id>\d+)")
-
-    def _get_filename_re(self) -> re.Pattern:
-        return re.compile(
-            r"(?P<name>.*)_(?P<well>[A-Z]+\d{2})_(?P<field>s\d+)_(?P<channel>w[1-9]{1})(?!_thumb)(?P<md_id>.*)(?P<ext>.tif)"
-        )
-
-    def _get_z_spacing(self) -> Optional[float]:
-        return None
