+++ src/faim_hcs/hcs/imagexpress/ImageXpressPlateAcquisition.py
@@ -1,0 +1,120 @@
+import os
+import re
+from abc import abstractmethod
+from pathlib import Path
+from typing import Optional, Union
+
+import pandas as pd
+
+from faim_hcs.hcs.acquisition import (
+    PlateAcquisition,
+    TileAlignmentOptions,
+    WellAcquisition,
+)
+from faim_hcs.hcs.imagexpress.ImageXpressWellAcquisition import (
+    ImageXpressWellAcquisition,
+)
+from faim_hcs.io.metadata import ChannelMetadata
+from faim_hcs.io.MetaSeriesTiff import load_metaseries_tiff_metadata
+
+
+class ImageXpressPlateAcquisition(PlateAcquisition):
+    def __init__(
+        self,
+        acquisition_dir: Union[Path, str],
+        alignment: TileAlignmentOptions,
+        background_correction_matrices: Optional[dict[str, Union[Path, str]]] = None,
+        illumination_correction_matrices: Optional[dict[str, Union[Path, str]]] = None,
+    ):
+        super().__init__(
+            acquisition_dir=acquisition_dir,
+            alignment=alignment,
+            background_correction_matrices=background_correction_matrices,
+            illumination_correction_matrices=illumination_correction_matrices,
+        )
+
+    def _parse_files(self) -> pd.DataFrame:
+        """Parse all files in the acquisition directory.
+
+        Returns
+        -------
+        DataFrame
+            Table of all files in the acquisition.
+        """
+        return pd.DataFrame(
+            self._list_and_match_files(
+                root_dir=self._acquisition_dir,
+                root_re=self._get_root_re(),
+                filename_re=self._get_filename_re(),
+            )
+        )
+
+    def _list_and_match_files(
+        self,
+        root_dir: Union[Path, str],
+        root_re: re.Pattern,
+        filename_re: re.Pattern,
+    ) -> list[str]:
+        files = []
+        for root, _, filenames in os.walk(root_dir):
+            m_root = root_re.fullmatch(root)
+            if m_root:
+                for f in filenames:
+                    m_filename = filename_re.fullmatch(f)
+                    if m_filename:
+                        row = m_root.groupdict()
+                        row |= m_filename.groupdict()
+                        row["path"] = str(Path(root).joinpath(f))
+                        files.append(row)
+        return files
+
+    @abstractmethod
+    def _get_root_re(self) -> re.Pattern:
+        """Regular expression for matching the root directory of the acquisition."""
+        raise NotImplementedError()
+
+    @abstractmethod
+    def _get_filename_re(self) -> re.Pattern:
+        """Regular expression for matching the filename of the acquisition."""
+        raise NotImplementedError()
+
+    def get_well_acquisitions(self) -> list[WellAcquisition]:
+        return [
+            ImageXpressWellAcquisition(
+                files=self._files[self._files["well"] == well],
+                alignment=self._alignment,
+                z_spacing=self._get_z_spacing(),
+                background_correction_matrices=self._background_correction_matrices,
+                illumination_correction_matrices=self._illumination_correction_matrices,
+            )
+            for well in self._files["well"].unique()
+        ]
+
+    @abstractmethod
+    def _get_z_spacing(self) -> Optional[float]:
+        raise NotImplementedError()
+
+    def get_channel_metadata(self) -> dict[str, ChannelMetadata]:
+        ch_metadata = {}
+        for ch in self._files["channel"].unique():
+            channel_files = self._files[self._files["channel"] == ch]
+            path = channel_files["path"].iloc[0]
+            metadata = load_metaseries_tiff_metadata(path=path)
+            from faim_hcs.MetaSeriesUtils import _build_ch_metadata
+
+            channel_metadata = _build_ch_metadata(metadata)
+            ch_metadata[ch] = ChannelMetadata(
+                channel_index=int(ch[1:]) - 1,
+                channel_name=ch,
+                display_color=channel_metadata["display-color"],
+                spatial_calibration_x=metadata["spatial-calibration-x"],
+                spatial_calibration_y=metadata["spatial-calibration-y"],
+                spatial_calibration_units=metadata["spatial-calibration-units"],
+                z_spacing=self._get_z_spacing(),
+                wavelength=channel_metadata["wavelength"],
+                exposure_time=channel_metadata["exposure-time"],
+                exposure_time_unit=channel_metadata["exposure-time-unit"],
+                objective=metadata["_MagSetting_"],
+            )
+
+        return ch_metadata
