+++ src/faim_hcs/hcs/imagexpress/__init__.py
@@ -1,3 +1,5 @@
+from .ImageXpressPlateAcquisition import ImageXpressPlateAcquisition  # noqa: F401
+from .ImageXpressWellAcquisition import ImageXpressWellAcquisition  # noqa: F401
 from .SinglePlaneAcquisition import SinglePlaneAcquisition  # noqa: F401
 from .StackAcquisition import StackAcquisition  # noqa: F401
 from .MixedAcquisition import MixedAcquisition  # noqa: F401
